requirements.txt
flask
openai
edge-tts
pillow
